Break down placement group tests
